#6 adding stylesheet for search panel and moving to component folder
