+++ app/components/SearchPanel.css
@@ -1,0 +1,8 @@
+.homeRow {
+  color: #844;
+}
+
+.rowVisited {
+  color: #616161;
+  font-style: italic;
+}
